#248 split tentacle manager
